requirements.txt
# Core
pytz==2021.1
Django==3.1.7
django-configurations==2.2
gunicorn==20.1.0

# For the persistence stores
psycopg2-binary==2.8.6
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.1.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==2.4.0

# Developer Tools
ipdb==0.13.7
ipython==7.22.0
mkdocs==1.1.2
flake8==3.9.0

# Testing
mock==4.0.3
factory-boy==3.2.0
pytest-cov==2.11.1
pytest-django==4.1.0
pytest==6.2.2
coverage==5.5
pytest-factoryboy==2.1.0
model-bakery==1.2.1

# Static and Media Storage
django-storages==1.11.1
boto3==1.17.43

